Remove openstack_handle service_names
Originally, the openstack refreshers had to look up each service,
and separately lookup each service_name to see if the service was the
appropriate service they wanted.

Now the openstack refreshers can rely on the returned service object to have contain the name of
the service directly.  This eliminates the need for the openstack refreshers to
carry around instance vars for both the services as well as the service names.


(transferred from ManageIQ/manageiq@9592011a8ec0545ba1e9c203a69b3596b6102be4)