Set OpenStack instance's cpu_speed
We need to determine an instance's cpu_speed because it is
used in the Right-Size Recommendation page.

Nova API doesn't provide a cpu_speed for user instances. There
is a cpu_speed for the baremetal host and we can use that as a
proxy. We gather this information during the cloud refresh
process.

The cpu_speed will be available when the fog-openstack
introspection provider merges.


(transferred from ManageIQ/manageiq@a015c9a1236c4dd27cc6fdf262e1d3996cb1afa8)