Filter networks and floating_ips for OpenStack provisioning
Filter networks and floating_ips for OpenStack provisioning. Only
non-external networks should be connectable to the VM. And only
floating IPs, that are connected through the router to the
chosen non external network should be available. Also we are
respecting the tenant associated to the floating ip. The
connection is now shown in the topology graph.


(transferred from ManageIQ/manageiq@44f91afdaff8ef9e75fadd87ea80f37d975b7b31)