Initial Checkin of StorageManager::SwiftManager
Add the model code for the new StorageManager::SwiftManager provider.
Note that this commit includes code that is redundant for the
StorageManager::CinderManager provider that is currently in a WIP PR at
https://github.com/ManageIQ/manageiq/pull/11088

A separate commit or commits is(are) still needed to implement UI changes.


(transferred from ManageIQ/manageiq@0f9a6586a1acc8c8b93d6e500e6cb4417ac74214)