Fix for Extra CloudObjectStoreContainer on Liberty
When support for CloudVolumeBackups is added in the Liberty release of Openstack the following
issue arises:
When volume backups are utilized, a CloudObjectStoreContainer named "volumes_backup" is created
automatically.  It is not visible in the Openstack Web UI.  Tests that don't expect this container
to exist will fail because it will throw off the expected count
This modification checks for this container and if so increases the number of expected
CloudObjectStoreContainers.


(transferred from ManageIQ/manageiq@3c8021fafe5ad11ecad80111768eeb486184633a)