Fix Gnocchi network metrics
Network metrics are missing on the overcloud. This happened when
we switched from Ceilometer to Gnocchi.

Unlike Ceilometer, Gnocchi network metrics are associated with the
instance's network interface and not the instance. This patch adds
additional logic to query for the network interface's resource
meters if they are available.

Fixes BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1434681