Only consider network interface when finding network metrics
Previously we were looking at all resources when we were finding
the network interface's resource id. We can reduce the number
of resources we inspect by limiting them to network interface
resources.

This requires a fog-openstack change to add a new request:
https://github.com/fog/fog-openstack/pull/283

This PR should be merged only when the fog-openstack change
has been released and the manageiq-gems-pending gemspec has been
updated with the release version containing that change.