Fix Ceilometer Event Monitor sync calls
MiqServer syncs processes each 15 seconds by default. This results into calling
available? method on event monitor. This was useful for AMQP subscriptions, but
it is not needed for Ceilometer and causes creating a lot of Keystone Auth token
on Openstack side.

This commit changes a method which is used for syncing workers to verify auth
status on provider for Ceilometer and keeps original behavior for AMQP.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1451122