Add fast refresh options
If ems_openstack.refresh.is_admin is set, then ManageIQ will attempt to use
admin-only api options to signficantly reduce the number of repeated API
calls necessary to build the Openstack inventory.

Due to differences in the way Heat's access control system works,
there is a separate flag to enable admin mode. This requires Heat to
be configured in a particular way.
See https://bugs.launchpad.net/heat/+bug/1466694 and
https://bugs.launchpad.net/keystone/+bug/968696 for more details
on why this is the case.