Use an unscoped connection to Keystone to validate API creds
`verify_api_credentials` was checking Openstack API credentials
against the Compute service without an explicit tenant name,
which lead to OpenstackHandle making a ton of connections to
decide on how to scope the request. As a result, verifying
Openstack credentials was extremely slow. Now it makes a single
request to Keystone, which is way faster.