Bypass the superclass orchestrated destroy for this Provider.
In the OpenStack provider, the Provider instance is only tightly coupled
to the InfraManager.  That is, when the InfraManager is created, then
the Provider is created and vice-versa, when the InfraManager is
destroyed, we need to destroy the Provider.  The CloudManager objects
associated to that Provider should *not* be destroyed, and only need to
be nullified.  For other Providers, the Provider instance is what is
destroyed, so by default the orchestrated destroy destroys all of its
managers, but we here don't want that to happen, so we are bypassing.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1538232