bz1573893: duplicate opts hash before modifying in raw_connect_try_ssl
https://bugzilla.redhat.com/show_bug.cgi?id=1573893

raw_connect_try_ssl modifies the opts hash (pulling out :ssl_options and then
using it. This means that if raw_connect_try_ssl is called twice with the
same opts hash, the second time fails because of the side effect of the first
call. With this change, the method works on a copy of the hash rather than
the original