Use nil as initialiser for @orchestration_stacks to prevent looping.
If there are no orchestration stacks in an environment then an empty list is returned. An empty
list causes .any? to return False and an empty list will always result in another query to Heat
the next time the method is called. As we make one query to Heat per tenant, this can quickly
add up. By using nil as the initialiser we separate the case of 'we have not yet fetched these
resources' from 'we fetched these resources and none were returned'. The former should allow
a refresh, the latter should not.

This is most pronounced in environments with a large number of cloud tenants and no orchestration
stacks, during the refresh of NetworkManager. There, the number of queries to Heat scales
according to the number of cloud networks and security groups in the environment, as the list of
orchestration stacks is fetched for each cloud network and security group. With incorrect
caching this can easily result in tens or hundreds of thousands of fruitless queries to Heat,
potentially taking the refresh time of NetworkManager from minutes to hours.