Manage snapshots as part of the inventory refresh process
Instance "snapshots" in Openstack are a collection of an image
and a set of volume snapshots. In Horizon these are managed
individually, but ManageIQ abstracts them into a "snapshot"
object with its own UI. These snapshots were not part of the
inventory refresh process and could get out of sync with
the Glance images they represented. This patch causes these
helper objects to be managed by the inventory refresh process.