Fixing targeted refresh for nested stacks and their resources
First the stack target was not extracted from the event, because
it goes as resource_id, not stack_id.

Second we need to always add all parent stacks as targets (so the
chain of relations gets connected) and we need to scan resources
to connect all instances to the stack.

(Issue found and fixed as part of hackaton)

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1732861