Update Network Targeted refresh to skip not found network
If an Event like network.delete.end is received, targeted refresh tries find the network
by a raw fog request method which doesn't handle Not Found error, that could break the refresh
attempt.

Updating to Fog's model method (network.get) which returns nil for Not Found error request
(similar approach as for other objects in targeted refresh).

This could be related to https://bugzilla.redhat.com/show_bug.cgi?id=1896581,
however we're not 100% sure whether this is the main problem on the BZ.