network_router changes to support the api
this is setup to work with both the old and the new ui.
I marked the differences in the ui so we could drop the old version.

the ui controllers was sending a cloud_tenant object over the wire
and was also doing the conversion from id to ref.

Now we are taking ids and converting them to to objecct and refs