Fix port for class-level verify_credentials
The class-level verify_credentials method takes DDF parameters with the
port being an integer.  The ems_connect? method was assuming the port
was a string so it was doing `params[:port].try(:strip)` which was
returning `nil` for an integer.  This then used the default port `5000`
since no port appeared to be passed.

The result is the class-level verify succeeds, then the instance-level
fails.