Allow a missing Snapshot#vm_or_template_id
When a snapshot is taken of a VM an image is created on the OpenStack
side.  We associate this as a snapshot record to the VM that created it.

If the VM is deleted then the lazy_find will return `nil` and a log or
raise an exception depending on prod/dev.

We can mark vm_or_template as optional in the manager_ref since the
snapshot uid is globally unique.