fix spec for supports_feature change
Since we will be no longer storing the `unsupported_reason` in a temporary variable,
calling `supports?()` and `unsupported_reason()` each check the `supports :feature` code.

So in this test, it calls it 2 times.

I removed the stub to allow the ems to be called multiple times without freaking a stub out
Alternative would have been to add to the ems stub: `at_least(1).times`