Rescue VM volume attachments NotFound
When a VM is in `ERROR` state (and possibly other states) the
vm.volume_attachments call may fail raising an
Fog::OpenStack::Compute::NotFound exception and causing the entire
inventory refresh to fail. This rescue allows the inventory refresh to
pass (and in the case of a VM in ERROR state, the VM will show up in
ManageIQ inventory as 'non_operational').